kernel/spec: remove the 512k offset
This offset used to be mandatory, but it is not any more. The gateware
I'm using has two aliases  in the 1M address space, so offset 0 is as good.
Other gateware (like the 100Msample ADC) use the whole address space,
so after this patch the ADC driver can use the fmc accessors like other
drivers.

EOF