kernel/wr-nic-dio: change of behaviour with periodic pulses
The period pulse on a channel was designed as a generic "mirror one channel
to another with a predefined delay".

However, on different boards self-stamping has different delays, so
if we are mirroring to ourselves, add the delay to the previous setpoint
instead of using the timestamp we got, so the 8ns difference doesn't
accumulate over time.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>