Makefiles: some contortions to be friends of submodules
This patch adds a script "check-fmc-bus" that returns the pathname of
the fmc-bus kernel driver, either under this directory or under
../fmc-bus .  The former applies when fmc-bus is a submodule of this
package, the latter applies when we both are submodules of another
package.

The final aim is helping the naive user, who just wants to build it
all in a single step -- both if building this package or another
package that depends on both fmc-bus and this one.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>