test-lm32: removed as obsolete stuff
This directory includes some tests to validate firmware loading
and other things, but they don't really apply to the new software layout.
Given things are pretty stable on that level, it's easier to remove
the directory (for which I had already dropped the docs) than to port
the code).

Thanks to Benoit Rat for noting this old code was still there.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>