general: upgraded to new fmc-bus version
This includes a new version of the submodule and fixes the code
to match the new set of fields in the fmc structure. The biggest change
is in the i2c address, that is now hosted in the fmc structure (this
changes the spec-i2c file).

Also, use new fmc_reprogram().  The function allows to centralize
dumping the eeprom and the sdb tree, so the bus framework dumps or
not, and can be changed at run time.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>