kernel/wr-nic (and doc): disable interrupts if too many
When using the SPEC as White Rabbit grandmaster, you need to feed it
with 10MHz, which would kill the computer with an interrupt flood.

DIO interrupts are now disabled if they consume more than 80% of the
total time, which usually triggers between 100kHz and 200kHz. Network
interrupts are not affected.

Thanks to Anders Wallin for reporting the problem and helping in the
diagnosis.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>