kernel: protect VIC on request,free and dispatch
Protecting the handlers' vector is not enough. Managing the interrupts
involves other tasks that must be serialized. Mainly to handle the first
request and the last free of a VIC handler.

The VIC internal locking (vec_lock) is not useful anymore, but I do not
remove it because theoretically the VIC should be an independent bunch of
code.

Signed-off-by: Federico Vaga <federico.vaga@cern.ch>