wr_nic: unify not runing the timer on SPEC
Removed if statement for SPEC is never reached due to return
in if(WR_IS_NODE) block.

Remove commit from spec-sw repo:
75780989f73d5b2ed154ba600e18d2936e302316
kernel/wr_nic/endpoint.c: don't run the timer

In the switch we used to have a timer to check for link status, but
in the SPEC card the internal lm32 does that, and access to the phy
is not atomic. Assume the link is always on until we'll get notification
from lm32 to the host.

Signed-off-by: Adam Wujek <adam.wujek@cern.ch>