wr-nic-gpio: do not check on gpiochip_remove
Since kernel 3.18 the gpiochip_remove function return void and not an
error code as before. In case the gpiochip is removed while there is a
requested gpio, the kernel will print an error message.

By applying this patch without any control on the kernel version we are
going to loose the error message in case of trouble with old kernels

Signed-off-by: Federico Vaga <federico.vaga@cern.ch>