kernel: swap irq ack between VIC and GN4124
We ack the GN4124 **before** the VIC. The other way around makes the
interrupt handling timing dependent. For instance. When the VIC
VIC_CTL_EMU_LEN_W is too short it may happens that the VIC asserts
the interrupt **before** the GN4124 is acked; then, so do not receive
interrupts anymore from the GN4124 because it misses the VIC edge.

Signed-off-by: Federico Vaga <federico.vaga@cern.ch>
Signed-off-by: Michel Arruat <michel.arruat@cern.ch>