Revert "kernel: fix memory leakage"
This reverts commit 0598de4086e0df5d9b9dbb6d5fbc83f90e477ed0.

This was not a memory leakage because the fmc-bus release that
pointer for us. Is it good?

We allocate and it is someone else us free, it lacks of simmetry.
Probably the register function should be called devm_ like all
the others in the kernel who relase for your the memory. In this
way it is clear the behaviour.