tools:vuart: do not use default offset value
There are no rule or specification that say that the VUART
offset for the White-rabbit core is at 0x20500. This is just a very
common value but it is not true in all cases.

Setting a default value based on fashion is not correct. Users
of bitstreams with other offsets expect the default to work because
they do not know what to do actually.

By removing the default we force the users to verify first where
the VUART is located.

Today SDB gives this information and once it is known, lazy people can
wrap the command in a script and pretending that it is the default.

Signed-off-by: Federico Vaga <federico.vaga@cern.ch>